+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### master
+
+<a name="v0.12.0"></a>
+### v0.12.0 - 2020-01-19
 - **breaking change:** commands given with `--cmd` must be separated (default separator is `;`)
 - fix some cases of terminal let in a bad state on errors (thanks Nathan West)
 - bring some changes to the fish shell function and its installation (PR #128)
@@ -6,6 +8,7 @@
 - don't use .gitignore files of parent repositories
 - change default value of the toggle_trim_root to false (fix #106 but might be reverted)
 - `:print_relative_path` verb (fix #169, thanks Roshan George)
+- `:chmod` verb
 
 <a name="v0.11.9"></a>
 ### v0.11.9 - 2020-01-15